Rename test/test.py to testunit.py, to avoid confusion with the root-level test.py
